refactor: add auth types file
